Version 2.0.0.0
I went completely away from using the built browser in prompts, and built a UI which integrates into the chat controls. I also switched from querying the close dialog for each question and instead querying the API one hundred questions at a time. 

If there are any returned questions that aren't closed, those questions are removed from the list. If the question does not show up in the response, it is either deleted or not a question (someone linked to an answer), so it stays in the list. It then shows the results and asks for your input to go further. 

At this point, you can either cancel (notice that the cancel button is available before this, but I wouldn't recommend clicking on it because I haven't written that part yet, otherwise it will just break stuff and make your browser do funky things) and it will forget everything and go back to the way it started off, or tell it to move the requests and it will move them in groups of 100. 

It's not perfect yet, but what ever is?